refactor(storage): extract dataclass records to records.py (Slice K-state-port phase 1, #1737)
Move the 15 @dataclass(slots=True) row shapes out of
src/pollypm/storage/state.py and into a new backend-agnostic module
src/pollypm/storage/records.py. The records are pure data shapes — no
sqlite logic — and are shared by both StateStore (sqlite, legacy) and
the per-table pg facades (work_session_queries, work_task_state,
memory_recall, etc.).

state.py keeps a re-export shim (``from .records import *`` via an
explicit list) so backwards-compatible ``from pollypm.storage.state
import SessionRecord`` keeps working until StateStore itself is
deleted in K-source-ripout.

Importers (51 files, src + tests) switched to
``from pollypm.storage.records import ...``. Files that still need
StateStore keep that import on a separate line and will be migrated
to pg facades in K-state-port phase 2.

No behaviour change — records are byte-identical (same field order,
defaults, slots=True).
