+++ src/pollypm/accounts.py
@@ -26,8 +26,6 @@ from pollypm.onboarding import (
 from pollypm.runtime_env import claude_config_dir, codex_home_dir, provider_profile_env
 from pollypm.session_services import create_tmux_client
 from pollypm.storage.state import StateStore
-
-
 @dataclass(slots=True)
 class AccountStatus:
     key: str
