+++ src/pollypm/launch_planner_protocol.py
@@ -36,8 +36,6 @@ from pollypm.providers.base import LaunchCommand
 if TYPE_CHECKING:
     from pollypm.config import PollyPMConfig
     from pollypm.storage.state import StateStore
-
-
 @dataclass(slots=True)
 class DefaultLaunchPlannerContext:
     """Callables the default planner needs from its host.
