+++ src/pollypm/llm_runner.py
@@ -21,7 +21,6 @@ from pollypm.config import PollyPMConfig, load_config
 from pollypm.models import ProviderKind
 from pollypm.runtime_env import claude_config_dir
 from pollypm.storage.state import StateStore
-
 logger = logging.getLogger(__name__)
 
 HAIKU_MODEL = "claude-haiku-4-5-20251001"
