+++ src/pollypm/transcript_ledger.py
@@ -8,9 +8,8 @@ from pathlib import Path
 from pollypm.config import load_config
 from pollypm.models import AccountConfig, ProviderKind
 from pollypm.projects import slugify_project_key
-from pollypm.storage.state import StateStore, TokenUsageHourlyRecord
-
-
+from pollypm.storage.records import TokenUsageHourlyRecord
+from pollypm.storage.state import StateStore
 @dataclass(slots=True)
 class TranscriptTokenSample:
     session_name: str
