+++ src/pollypm/work/cli.py
@@ -2167,7 +2167,6 @@ def task_pickup_log(
     try:
         from pollypm.config import DEFAULT_CONFIG_PATH, load_config, resolve_config_path
         from pollypm.storage.state import StateStore
-
         config_path = resolve_config_path(DEFAULT_CONFIG_PATH)
         if not config_path.exists():
             typer.echo(
