+++ tests/integration/test_checkpoint_levels_integration.py
@@ -24,8 +24,6 @@ from pollypm.models import (
     SessionLaunchSpec,
 )
 from pollypm.storage.state import StateStore
-
-
 def _config(tmp_path: Path) -> PollyPMConfig:
     project_root = tmp_path / "repo"
     project_root.mkdir()
