+++ tests/test_architect_lifecycle.py
@@ -28,8 +28,6 @@ from pollypm.models import ProviderKind
 from pollypm.providers.claude import ClaudeProvider
 from pollypm.providers.codex import CodexProvider
 from pollypm.storage.state import StateStore
-
-
 @pytest.fixture
 def store(tmp_path: Path) -> StateStore:
     return StateStore(tmp_path / "state.db")
